Corrected visibility of `Ext` traits
This addresses #93 and builds upon the previous commits to this pull request.

Bear in mind that I have no idea what I'm doing—I've had to manually edit many automatically generated files throughout this PR (which I know I shouldn't have to be doing) to get this to work, but, hey, it works!

There are now eight warnings, and no errors.

I'm not sure if this can be merged just yet, for reasons such as code style and the whatnot; a review is still required